fixed a bug for item edit

+++ Project/views/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, HiddenField, TextAreaField, IntegerField, URLField, SubmitField
+from wtforms import StringField, HiddenField, TextAreaField, IntegerField, URLField, SubmitField, FloatField
 from wtforms.validators import DataRequired, Length, Optional, NumberRange
 # Gagan Indukala Krishna Murthy - gi36 - April 21th
 class ItemForm(FlaskForm):
@@ -7,6 +7,6 @@ class ItemForm(FlaskForm):
     name = StringField("name", validators=[DataRequired(), Length(max=30)])
     description = TextAreaField("description", validators=[DataRequired()])
     stock = IntegerField("stock", validators=[NumberRange(min=0)])
-    unit_price = IntegerField("unit_price", validators=[NumberRange(min=0)])
+    unit_price = FloatField("unit_price", validators=[NumberRange(min=0)])
     image = URLField("image", validators=[Optional()])
     submit = SubmitField("Save")
